+++ lib/models/points-prelevement.js
@@ -2,7 +2,7 @@ import {chain, minBy, uniq} from 'lodash-es'
 
 import mongo from '../util/mongo.js'
 import {isAfter} from 'date-fns'
-import {validateCreation} from '../validation/points-validation.js'
+import {validateCreation} from '../validation/point-validation.js'
 import {nanoid} from 'nanoid'
 
 export async function decoratePointPrelevement(point) {
