+++ lib/validation/point-validation.js
